PoissonSolvers Library Functions

PoissonSolvers.PoissonSolution — Function

Takes a basis and a coefficient vector and returns an appropriate object that allows to evaluate the corresponding solution via a functor via

sol = PoissonSolution(basis, coeffs)
sol(x)

PoissonSolvers.solve! — Function

Solves the Poisson equation for periodic boundary conditions

function solve!(p::PoissonSolver, x::AbstractVector, w::AbstractVector) end